LayoutTests/fast/selectors/ignore-pseudo-element-inside-non-rightmost-fragments-expected.txt

Ignore pseudo elements inside non-rightmost fragments when collecting rules for a node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getMatchedCSSRules(document.getElementById('testcase1'), '') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase2'), '') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase1'), '::after') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase2'), '::after') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase1'), '::before') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase2'), '::before') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase1'), '::first-letter') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase2'), '::first-letter') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase1'), '::first-line') is null
PASS window.getMatchedCSSRules(document.getElementById('testcase2'), '::first-line') is null
PASS successfullyParsed is true

TEST COMPLETE
